update git extraction
